откатился на момент когда не написал сервер и клиент
