--- .idea/artifacts/Notes_jar.xml
@@ -1,29 +1,0 @@
-<component name="ArtifactManager">
-  <artifact build-on-make="true" name="Notes:jar">
-    <output-path>$PROJECT_DIR$/out/artifacts/Notes_jar</output-path>
-    <root id="root">
-      <element id="archive" name="Notes.jar">
-        <element id="module-output" name="Notes" />
-      </element>
-      <element id="library" level="project" name="Maven: org.openjfx:javafx-fxml:17.0.1" />
-      <element id="library" level="project" name="Maven: org.openjfx:javafx-base:17.0.1" />
-      <element id="library" level="project" name="Maven: org.kordamp.ikonli:ikonli-javafx:12.2.0" />
-      <element id="library" level="project" name="Maven: com.dlsc.formsfx:formsfx-core:11.3.2" />
-      <element id="library" level="project" name="Maven: org.controlsfx:controlsfx:11.1.0" />
-      <element id="library" level="project" name="Maven: net.synedra:validatorfx:0.1.13" />
-      <element id="library" level="project" name="Maven: org.openjfx:javafx-graphics:win:17.0.1" />
-      <element id="library" level="project" name="Maven: org.openjfx:javafx-fxml:win:17.0.1" />
-      <element id="library" level="project" name="Maven: org.kordamp.bootstrapfx:bootstrapfx-core:0.4.0" />
-      <element id="library" level="project" name="Maven: org.kordamp.ikonli:ikonli-core:12.2.0" />
-      <element id="library" level="project" name="Maven: org.openjfx:javafx-controls:win:17.0.1" />
-      <element id="library" level="project" name="Maven: org.openjfx:javafx-web:win:17.0.1" />
-      <element id="library" level="project" name="Maven: org.openjfx:javafx-media:win:17.0.1" />
-      <element id="library" level="project" name="Maven: org.openjfx:javafx-graphics:17.0.1" />
-      <element id="library" level="project" name="Maven: org.openjfx:javafx-base:win:17.0.1" />
-      <element id="library" level="project" name="Maven: org.openjfx:javafx-controls:17.0.1" />
-      <element id="library" level="project" name="Maven: org.openjfx:javafx-media:17.0.1" />
-      <element id="library" level="project" name="Maven: org.openjfx:javafx-web:17.0.1" />
-      <element id="library" level="project" name="Maven: eu.hansolo:tilesfx:11.48" />
-    </root>
-  </artifact>
-</component>
