--- .idea/libraries/Maven__org_openjfx_javafx_graphics_linux_17_0_1.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Maven: org.openjfx:javafx-graphics:linux:17.0.1">
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/openjfx/javafx-graphics/17.0.1/javafx-graphics-17.0.1-linux.jar!/" />
-    </CLASSES>
-    <JAVADOC>
-      <root url="jar://$MAVEN_REPOSITORY$/org/openjfx/javafx-graphics/17.0.1/javafx-graphics-17.0.1-javadoc.jar!/" />
-    </JAVADOC>
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/openjfx/javafx-graphics/17.0.1/javafx-graphics-17.0.1-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
