--- META-INF/MANIFEST.MF
@@ -1,4 +1,0 @@
-Manifest-Version: 1.0
-Main-Class: com.example.notes.Main
-
-
