+++ src/main/java/module-info.java
@@ -1,14 +1,11 @@
-module com.example.notes {
+module com.example.notes  {
     requires javafx.controls;
     requires javafx.fxml;
+
     requires com.dlsc.formsfx;
     requires org.kordamp.ikonli.javafx;
+    requires javafx.graphics;
 
+    opens com.example.notes to javafx.fxml;
     exports com.example.notes;
-    exports com.example.notes.Remote;
-
-    opens com.example.notes to
-            javafx.fxml;
-    opens com.example.notes.Remote to
-            javafx.fxml;
 }
