--- src/main/resources/META-INF/MANIFEST.MF
@@ -1,10 +1,0 @@
-Manifest-Version: 1.0
-Main-Class: com.example.notes.Main
-Class-Path: javafx-fxml-17.0.1.jar javafx-base-17.0.1.jar ikonli-javafx-
- 12.2.0.jar formsfx-core-11.3.2.jar controlsfx-11.1.0.jar validatorfx-0.
- 1.13.jar javafx-graphics-17.0.1-win.jar javafx-fxml-17.0.1-win.jar boot
- strapfx-core-0.4.0.jar ikonli-core-12.2.0.jar javafx-controls-17.0.1-wi
- n.jar javafx-web-17.0.1-win.jar javafx-media-17.0.1-win.jar javafx-grap
- hics-17.0.1.jar javafx-base-17.0.1-win.jar javafx-controls-17.0.1.jar j
- avafx-media-17.0.1.jar javafx-web-17.0.1.jar tilesfx-11.48.jar
-
