+++ src/main/resources/com/example/notes/Notes.txt
@@ -1,5 +1,4 @@
 первый
 второй
 третий
-dfggdfg
-Стартуем
+всякое
