+++ src/main/resources/com/example/notes/hello-view.fxml
@@ -46,9 +46,9 @@
         </MenuBar>
     </top>
     <center>
-        <ScrollPane fx:id="scrollPanewithVbox" fitToHeight="true" fitToWidth="true" prefHeight="285.0" prefWidth="376.0" BorderPane.alignment="CENTER">
+        <ScrollPane fx:id="scrollPanewithVbox" fitToWidth="true" BorderPane.alignment="CENTER">
             <content>
-                <VBox fx:id="vboxForButtons" accessibleRole="NODE" prefHeight="254.0" prefWidth="374.0" />
+                <VBox fx:id="vboxForButtons" />
             </content>
         </ScrollPane>
     </center>
