+++ src/main/resources/com/example/notes/style_hello_view.css
@@ -3,7 +3,7 @@
   -fx-text-fill: white;
   -fx-background-insets:3px;
   -fx-border-raduis:5px;
-
+  
   }
   .root{
   -fx-background-color: black;
